Removed overage, added shared budgeting
Overage was just a weird concept. Ultimately the same thing can be
achieved by setting a flag in your own application and running with
usize::MAX. To simplify things, I removed it.

Shared budgeting is now possible with ReplenishableBudget. While the
current tests show how it can be used to drive a single task, it can be
used to drive multiple tasks using a shared budget.