Introduced threadsafe and singlethreaded
This was a doozie of a refactor, but thankfully Rust is a charm at
refactoring. The net result is four total surface APIs. One for each
async and non-async, and for each of those, one that is Send + Sync and
one that is !Send and !Sync.

While it's a lot of API surface, it means that the end-user picks which
implementation they want by just importing the correct module.