Switched threadsafe blocking to use condvar
There was a race condition where a thread could replenish budget between
the budget being denied and the thread being parked. When this happened,
the parked thread would never wake up in the unit test, since the budget
is only replenished when messages are received by the task that spends
the budget.

Technically, I could have avoided the condvar, but it would have
required an additional atomic anyways -- so it seemed like a clear win
to simplify and just use Mutex/Condvar for the sleeping too.

Now whenever budget is denied, the generation is recorded. When
preparing to park, if the generation is changed, we skip parking to
allow the futures to run again.