Fixed more potential deadlocks
Added a new unit test that stresses the locking behavior.

The main deadlock found was that by consuming the waker I was
introducing a race condition between when the waker was registered and
the runtime placed it in the pending queue. If a replenish call was
issued between that time, the future would end up in the pending queue
and never come back.

Because add/remove_waker are paired, switching to waking by ref ensures
that a future replenish call will wake up any still sleeping futures.

This does mean that there is one final edge case for the nightmare()
unit test, but because of the thread sleeping, I can't get it to fail.
If the final task lands in that window and there is never another
replenish call, the future won't wake up.

Splitting the queues and keeping the pending queue locked while the
runtime is executing is an option, but I think I'm happy with this for
now as it's not a common use case -- I envision ReplenishbleBudget being
used to replenish budgets on a time basis, ensuring scripts can't get
too greedy. In those situations, there will always be another replenish
call.